Moved SCPG pass to it's own pass file
