+++ plume/src/test/kotlin/io/github/plume/oss/drivers/NeptuneDriverIntTest.kt
@@ -43,6 +43,7 @@ import overflowdb.Graph
 import scala.Option
 import kotlin.properties.Delegates
 
+@Suppress("DEPRECATION")
 class NeptuneDriverIntTest {
 
     companion object {
